New intl_* page cyclers are crashing on perf bots
The new 8 bit KURL code was missing a null check. The version of this code in
KURL already has this check, but this hand-rolled URL parsing code was missing
it. This code looks a lot like premature optimization. Given all the other work
that this code is doing, I'm surpised the overhead of creating a KURL is
noticable, but that's a CL for another day.

BUG=241158

Review URL: https://codereview.chromium.org/14632013

git-svn-id: svn://svn.chromium.org/blink/trunk@150454 bbb929c8-8fbe-4397-9dbb-9b2b20218538