Remove TypeCanFailConversion() from CodeGeneratorV8.pm
TypeCanFailConversion() is just for "Attr" type arguments, which are used
in a couple of methods in Element.idl (setAttributeNode(), removeAttributeNode(),
setAttributeNodeNS()). This should be replaced with [StrictTypeChecking].
The only difference is the exception name. [StrictTypeChecking] throws a TypeError
whereas TypeCanFailConversion() throws TYPE_MISMATCH_ERR. Given that those three methods
are already deprecated in the spec, a slight change in the exception name won't
cause a compatibility problem.

R=adamk@chromium.org

Review URL: https://codereview.chromium.org/14898010

git-svn-id: svn://svn.chromium.org/blink/trunk@150469 bbb929c8-8fbe-4397-9dbb-9b2b20218538