Improve shadow support in EllipsisBox.
Previously, the ellipsis generated by text-overflow: ellipsis would not
render correctly if multiple text-shadows were specified, or if the text
color was not opaque (but there was a shadow).

This brings the behaviour of EllipsisBox more in line with InlineTextBox (as far as text-shadow is concerned).

BUG=241073

Review URL: https://chromiumcodereview.appspot.com/15001034

git-svn-id: svn://svn.chromium.org/blink/trunk@150534 bbb929c8-8fbe-4397-9dbb-9b2b20218538