Fix flakiness dashboard loader to handle failure_map top-level key.
This code assumed that the builderName and the version were the only
top-level keys. Also, it's super outdated. We pass in the builderName now,
no need to try and get it out of the JSON. Long-term we should change
the JSON format to have a top-level builder_name key and move all the 
keys currently under the builderName up one level.

Also, remove the version check. All the JSON has been version 4 for a long time.

NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/15311002

git-svn-id: svn://svn.chromium.org/blink/trunk@150604 bbb929c8-8fbe-4397-9dbb-9b2b20218538