Include bug entries in the results json.
This is so we can show bugs in the flakiness dashboard.
The bulk of this change is making it so that we stop storing the legacy
BUGXXX format in test_expectations and actually store the direct entry
from the TestExpectations line. We should move all this code over
to using the new syntax, but doing it for the bug entries is a good first
step.

Review URL: https://chromiumcodereview.appspot.com/15422003

git-svn-id: svn://svn.chromium.org/blink/trunk@150711 bbb929c8-8fbe-4397-9dbb-9b2b20218538