Don't clear the renderer in Document::attach
There doesn't appear to be any reason to clear the Document's renderer
before calling the super class ContainerNode::attach except for the
assert in ContainerNode::attach which is easily fixed.

This code looks to be more than 7 years old and probably dates back to
a different world.

Review URL: https://chromiumcodereview.appspot.com/15408003

git-svn-id: svn://svn.chromium.org/blink/trunk@150750 bbb929c8-8fbe-4397-9dbb-9b2b20218538