A RenderLayer shouldn't report scroll range if it doesn't scroll
This fixes a regression introduced in r150264, that scrolling a frame feels
choppy with pages which use <html style="height:100%">.

The cause of this bug is that both FrameView and the RenderLayer owned by the
RenderView reports scroll range, however the RenderView doesn't actually
handle scrolling. Therefore some of the scroll delta get absorbed by the
RenderView.

BUG=242113
R=jchaffraix@chromium.org

Review URL: https://codereview.chromium.org/15663002

git-svn-id: svn://svn.chromium.org/blink/trunk@150938 bbb929c8-8fbe-4397-9dbb-9b2b20218538