[CSSRegions] Moving to a flow thread should not overwrite rules of renderer creation
When moving an element to a flow thread we should take into account the rules used
to decide whether a renderer should be created for the element in the first place.
If without moving the element to a flow thread, the element would not get a renderer,
by moving to a flow thread, the same rules should be preserved.

BUG=242416
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/15752002

git-svn-id: svn://svn.chromium.org/blink/trunk@151256 bbb929c8-8fbe-4397-9dbb-9b2b20218538