DevTools: [CodeMirror] Do not force selected text color under highlighted ranges
Prelude: CodeMirror's selection is rendered with spans and thus has nothing in common with
native selection color. To make it look as if its a real selection we
preset its foreground selection color and background selection color into
renderer's selection colors.

This patch fixes a specific case with text ranges which are highlighted with TextEditor.highlightRange method
and which are selected at the same time. These ranges are rendred above selection and most
of the time override text's background color. Sadly, most of the time this results in unreadable text as it has
white foreground and light background.

This patch doesn't preset selected text foreground color if the text is a part of a highlighted text range.

BUG=
R=vsevik@chromium.org

Review URL: https://codereview.chromium.org/15804008

git-svn-id: svn://svn.chromium.org/blink/trunk@151393 bbb929c8-8fbe-4397-9dbb-9b2b20218538